# Avoid allocations in ID conversions