**What this PR does**:
Small change in ID conversion that avoids an allocation. Since this is called in the `spanCollector` and `traceCollecor` this has a significant impact on allocations for metrics queries.

**Which issue(s) this PR fixes**:
None

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`