@stoewer 2024-07-02
Benchmark for metrics:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: AMD Ryzen 7 PRO 6850U with Radeon Graphics     
                                                                             │ bench-metrics-03-main.txt │   bench-metrics-03-id-alloc.txt    │
                                                                             │          sec/op           │    sec/op     vs base              │
BackendBlockQueryRange/{}_|_rate()/7-16                                                      2.172 ± 21%    2.102 ± 17%       ~ (p=0.234 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                            3.781 ± 13%    3.546 ± 10%       ~ (p=0.382 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                           2.623 ± 13%    2.477 ± 13%       ~ (p=0.382 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                   4.207 ±  3%    4.084 ±  4%  -2.93% (p=0.038 n=8)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                232.0m ±  8%   223.3m ±  4%  -3.71% (p=0.050 n=8)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                         290.4m ±  1%   287.3m ±  1%  -1.08% (p=0.028 n=8)
geomean                                                                                      1.352          1.301        -3.80%

                                                                             │ bench-metrics-03-main.txt │    bench-metrics-03-id-alloc.txt    │
                                                                             │         allocs/op         │  allocs/op    vs base               │
BackendBlockQueryRange/{}_|_rate()/7-16                                                     522.7k ±  0%   363.3k ±  0%  -30.49% (p=0.000 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                           553.3k ±  0%   394.0k ±  0%  -28.80% (p=0.007 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                          524.7k ±  0%   365.3k ±  0%  -30.38% (p=0.007 n=8)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                  2.623M ± 19%   2.309M ± 15%        ~ (p=0.065 n=8)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                117.9k ±  0%   109.8k ±  0%   -6.82% (p=0.000 n=8)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                         76.55k ±  0%   71.03k ±  0%   -7.22% (p=0.000 n=8)
geomean                                                                                     391.3k         313.1k        -19.99%

```

TraceQL
```
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: AMD Ryzen 7 PRO 6850U with Radeon Graphics     
                                                   │ bench-traceql-03-main.txt │   bench-traceql-03-id-alloc.txt   │
                                                   │          sec/op           │   sec/op     vs base              │
BackendBlockTraceQL/spanAttValMatch-16                             553.3m ± 0%   539.1m ± 0%  -2.56% (p=0.000 n=8)
BackendBlockTraceQL/spanAttValNoMatch-16                           75.25m ± 1%   74.58m ± 1%  -0.88% (p=0.005 n=8)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                       164.4m ± 2%   162.7m ± 1%  -1.02% (p=0.015 n=8)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                     4.025m ± 1%   3.995m ± 0%       ~ (p=0.083 n=8)
BackendBlockTraceQL/resourceAttValMatch-16                          1.820 ± 1%    1.787 ± 0%  -1.81% (p=0.000 n=8)
BackendBlockTraceQL/resourceAttValNoMatch-16                       3.780m ± 1%   3.693m ± 1%  -2.29% (p=0.000 n=8)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                   342.2m ± 1%   328.9m ± 1%  -3.89% (p=0.000 n=8)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                 3.866m ± 1%   3.730m ± 0%  -3.52% (p=0.000 n=8)
BackendBlockTraceQL/mixedValNoMatch-16                             663.9m ± 3%   653.3m ± 1%       ~ (p=0.105 n=8)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                       52.73m ± 1%   51.67m ± 1%  -2.02% (p=0.000 n=8)
BackendBlockTraceQL/mixedValMixedMatchOr-16                        538.6m ± 4%   539.2m ± 1%       ~ (p=0.382 n=8)
BackendBlockTraceQL/count-16                                        1.460 ± 0%    1.452 ± 2%       ~ (p=0.505 n=8)
BackendBlockTraceQL/struct-16                                       2.547 ± 3%    2.553 ± 5%       ~ (p=0.645 n=8)
BackendBlockTraceQL/||-16                                          565.1m ± 1%   565.9m ± 2%       ~ (p=0.798 n=8)
BackendBlockTraceQL/mixed-16                                       211.7m ± 1%   212.2m ± 1%       ~ (p=0.645 n=8)
BackendBlockTraceQL/complex-16                                     4.387m ± 2%   4.317m ± 1%  -1.60% (p=0.021 n=8)
geomean                                                            132.7m        130.9m       -1.37%

                                                   │ bench-traceql-03-main.txt │   bench-traceql-03-id-alloc.txt   │
                                                   │         allocs/op         │  allocs/op   vs base              │
BackendBlockTraceQL/spanAttValMatch-16                             2.174M ± 0%   2.004M ± 0%  -7.82% (p=0.000 n=8)
BackendBlockTraceQL/spanAttValNoMatch-16                           42.03k ± 0%   42.03k ± 0%       ~ (p=0.287 n=8)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                       207.8k ± 0%   202.6k ± 0%  -2.54% (p=0.000 n=8)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                     41.69k ± 0%   41.69k ± 0%       ~ (p=0.706 n=8)
BackendBlockTraceQL/resourceAttValMatch-16                        10.039M ± 0%   9.856M ± 0%  -1.82% (p=0.000 n=8)
BackendBlockTraceQL/resourceAttValNoMatch-16                       41.72k ± 0%   41.72k ± 0%       ~ (p=0.608 n=8)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                   1.360M ± 0%   1.250M ± 0%  -8.08% (p=0.000 n=8)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                 41.72k ± 0%   41.72k ± 0%       ~ (p=0.695 n=8)
BackendBlockTraceQL/mixedValNoMatch-16                             46.75k ± 0%   46.84k ± 0%       ~ (p=0.278 n=8)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                       43.72k ± 0%   43.72k ± 0%  -0.01% (p=0.000 n=8)
BackendBlockTraceQL/mixedValMixedMatchOr-16                        44.71k ± 0%   44.70k ± 0%       ~ (p=0.768 n=8)
BackendBlockTraceQL/count-16                                       5.982M ± 0%   5.802M ± 0%  -3.01% (p=0.000 n=8)
BackendBlockTraceQL/struct-16                                      4.391M ± 0%   4.403M ± 1%       ~ (p=0.878 n=8)
BackendBlockTraceQL/||-16                                          207.1k ± 0%   195.2k ± 0%  -5.74% (p=0.000 n=8)
BackendBlockTraceQL/mixed-16                                       214.3k ± 0%   205.4k ± 0%  -4.15% (p=0.000 n=8)
BackendBlockTraceQL/complex-16                                     42.11k ± 0%   42.11k ± 0%       ~ (p=0.394 n=8)
geomean                                                            234.6k        229.7k       -2.09%

```